Avoid warning in VS 2022
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@6460 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COpenGLDriver.cpp
@@ -804,7 +804,7 @@ IRenderTarget* COpenGLDriver::addRenderTarget()
 
 
 // small helper function to create vertex buffer object address offsets
-static inline const GLvoid * buffer_offset(const long offset)
+static inline const GLvoid * buffer_offset(const size_t offset)
 {
 	return (const GLvoid *)offset;
 }
